chore: format, structure, clean up
