--- src/client/providers/settings/settings.ts
@@ -1,8 +1,0 @@
-import { UserSettings } from "../../types.js";
-
-export const settings: UserSettings = {
-    userId: 1,
-    settings: {
-        toolApprovalRequired: true,
-    },
-}
