+++ src/client/types.ts
@@ -42,47 +42,3 @@ export interface IMCPClientWrapper extends ToolProvider {
     // listResources(): Promise<string[]>;
     // readResource(url: string): Promise<string>;
 }
-
-/**
- * Interface for tool confirmation provider for client
- */
-export interface UserConfirmationProvider {
-    allowedTools: Set<string>,
-    requestConfirmation(
-        details: ToolExecutionDetails,
-        callbacks?: {
-            displayDetails?: (details: ToolExecutionDetails) => void;
-            collectInput?: () => Promise<string | boolean>;
-            parseResponse?: (response: any) => boolean;
-        },
-    ): Promise<boolean>;
-}
-
-/**
- * Interface for current tool being executed
- */
-export interface ToolExecutionDetails {
-    toolName: string;
-    args: any;
-    description?: string;
-}
-
-/**
- * Interface for user settings
- * TODO: This is a placeholder and should be replaced with a proper schema
- */
-export interface UserSettings {
-    userId?: number,
-    settings: {
-        [key: string]: any;
-    };
-}
-
-/**
- * Interface for settings provider
- * TODO: This is a placeholder and should be replaced with a proper schema
- */
-export interface SettingsProvider {
-    getUserSettings(userId?: string): Promise<UserSettings>;
-    updateUserSettings(userId: string, settings: UserSettings): Promise<void>;
-}
